Lista MetodiDiPagamento

+++ src/main/java/it/univaq/disim/oop/roc/business/impl/file/FileMetodiServiceImpl.java
@@ -20,7 +20,7 @@ public class FileMetodiServiceImpl implements MetodiService {
 	private static final String METODI_FILE_NAME = REPOSITORY_BASE + File.separator + "metodi.txt";
 
 	@Override
-	public Carta addCarta(Carta carta, String nomeCarta, String intestatario, Integer numero, Integer meseScadenza,
+	public Carta addCarta(Carta carta, String nomeCarta, String intestatario, Long numero, Integer meseScadenza,
 			Integer annoScadenza, Integer cvv) throws BusinessException {
 		try {
 			FileData fileData = Utility.readAllRows(METODI_FILE_NAME);
